LRCI-3965 Create BaseGitHubCommentEvent to be a base for all Comment type event handlers
