+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/event/github/BaseGitHubCommentEventHandler.java
@@ -1,0 +1,41 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.jethr0.event.github;
+
+import com.liferay.jethr0.event.EventHandlerContext;
+import com.liferay.jethr0.event.github.comment.GitHubComment;
+
+import org.json.JSONObject;
+
+/**
+ * @author Michael Hashimoto
+ */
+public abstract class BaseGitHubCommentEventHandler
+	extends BaseGitHubIssueEventHandler {
+
+	protected BaseGitHubCommentEventHandler(
+		EventHandlerContext eventHandlerContext, JSONObject messageJSONObject) {
+
+		super(eventHandlerContext, messageJSONObject);
+	}
+
+	protected GitHubComment getGitHubComment() throws InvalidJSONException {
+		JSONObject messageJSONObject = getMessageJSONObject();
+
+		JSONObject commentJSONObject = messageJSONObject.optJSONObject(
+			"comment");
+
+		if (commentJSONObject == null) {
+			throw new InvalidJSONException(
+				"Missing \"comment\" from message JSON");
+		}
+
+		GitHubFactory gitHubFactory = getGitHubFactory();
+
+		return gitHubFactory.newGitHubComment(commentJSONObject);
+	}
+
+}
